backend/requirements.txt
absl-py==2.1.0
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
binance-connector==3.6.0
ccxt==4.0.101
celery==5.3.6
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
contourpy==1.1.1
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
#cryptography==41.0.4
cryptography==42.0.7
cycler==0.11.0
daphne==4.1.0
defusedxml==0.8.0rc2
Django==5.0.2
django-celery-beat==2.6.0
django-channels==0.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
django-two-factor-auth==1.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
flatbuffers==24.3.25
fonttools==4.42.1
frozenlist==1.4.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itypes==1.2.0
Jinja2==3.1.3
joblib==1.3.2
keras==3.3.3
kiwisolver==1.4.5
kombu==5.3.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpld3==0.5.10
msgpack==1.0.8
multidict==6.0.4
namex==0.0.8
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.1.0
phonenumbers==8.13.30
Pillow==10.0.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycares==4.3.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
pypng==0.20220715.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.3
regex==2024.5.10
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.20.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.2.0
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.3
whitenoise==6.6.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.2
